animations/requirements.txt
websocket-client
numpy
matplotlib
aiohttp
sounddevice
